refactor(titlebar): move plus button to statusbar, increase WMUX font size
Move the add-workspace button from titlebar left segment to statusbar right side.
Increase WMUX wordmark font size from text-xs to text-sm for better visibility.
PresetPicker is now rendered at the end of StatusBar for proper positioning.
